This is a program that runs a maze game
This program is really just a maze game implemented into a menu framework/template I've been working on to make building complex programs easier to make.
This template takes a lot of the legwork out of dealing with creating utilities and a bunch of functions to prop up a program and dialog within console menu-based themes.
I update this template semi-regularly and my main focus is introducing file I/O into the capabilities of the framework. This current version has an assignment I've been working
on for class implemented into it.